+++ packages/gesturehandler/platforms/ios/src/FlingHandler.h
@@ -1,6 +1,6 @@
 #import "GestureHandler.h"
 
 @interface FlingGestureHandler : GestureHandler
-//@property (nonatomic) NSNumber * direction;
-//@property (nonatomic) NSNumber * numberOfTouchesRequired;
+- (void) setDirection:(NSInteger) value;
+- (void) setNumberOfTouchesRequired:(NSInteger) value;
 @end