+++ packages/gesturehandler/platforms/ios/src/LongPressHandler.h
@@ -9,4 +9,7 @@
 #import "GestureHandler.h"
 
 @interface LongPressGestureHandler : GestureHandler
+
+- (void) setMinDurationMs:(CGFloat) value;
+- (void) setMaxDist:(CGFloat) value;
 @end