+++ packages/gesturehandler/platforms/ios/src/PanHandler.h
@@ -10,5 +10,29 @@
 
 @interface PanGestureHandler : GestureHandler
 
+//@property (nonatomic) CGFloat minDist;
+//@property (nonatomic) CGFloat minVelocityX;
+//@property (nonatomic) CGFloat minVelocityY;
+//@property (nonatomic) CGFloat activeOffsetXStart;
+//@property (nonatomic) CGFloat activeOffsetXEnd;
+//@property (nonatomic) CGFloat failOffsetXStart;
+//@property (nonatomic) CGFloat failOffsetXEnd;
+//@property (nonatomic) CGFloat activeOffsetYStart;
+//@property (nonatomic) CGFloat activeOffsetYEnd;
+//@property (nonatomic) CGFloat failOffsetYStart;
+//@property (nonatomic) CGFloat failOffsetYEnd;
+
+- (void)setMinDist:(CGFloat)value;
+- (void)setMinVelocityX:(CGFloat)value;
+- (void)setMinVelocityY:(CGFloat)value;
+- (void)setMinVelocitySq:(CGFloat)value;
+- (void)setActiveOffsetXStart:(CGFloat)value;
+- (void)setActiveOffsetXEnd:(CGFloat)value;
+- (void)setFailOffsetXStart:(CGFloat)value;
+- (void)setFailOffsetXEnd:(CGFloat)value;
+- (void)setActiveOffsetYStart:(CGFloat)value;
+- (void)setActiveOffsetYEnd:(CGFloat)value;
+- (void)setFailOffsetYStart:(CGFloat)value;
+- (void)setFailOffsetYEnd:(CGFloat)value;
 
 @end